Se agrego un mensaje de advertencia cuando valla a limpiar todas las tareas completadas

+++ public/js/app.js
@@ -77,10 +77,13 @@ async function deleteTask(id) {
   renderList();
 }
 async function clearCompleted() {
-  const completed = allTasks.filter(t => t.completed);
-  await Promise.all(completed.map(t => fetch(`/api/tasks/${t.id}`, { method: 'DELETE' })));
-  allTasks = allTasks.filter(t=>!t.completed);
-  renderList();
+  const confirmDelete = confirm("¿Está seguro? Se borrarán todas las tareas completadas.");
+    if (confirmDelete) {
+        const completed = allTasks.filter(t => t.completed);
+        await Promise.all(completed.map(t => fetch(`/api/tasks/${t.id}`, { method: 'DELETE' })));
+        allTasks = allTasks.filter(t=>!t.completed);
+        renderList();
+    }
 }
 
 // --- Render ---
